requirements.txt
fastapi
pydantic
uvicorn
requests
openai
python-multipart
python-dotenv